requirements.txt
feedparser>=6.0.11
playwright>=1.40.0
pyyaml>=6.0
python-dateutil>=2.8
openai>=1.40.0
